added measuring units

+++ SET3-Backend/Models/OrderModel.cs
@@ -9,10 +9,10 @@ namespace SET3_Backend.Models
         public int Id { get; set; }
         public int ShopId { get; set; }
         public DateTime Date { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public int ProductId { get; set; }
 
-        public OrderModel(int shopId, DateTime date, int quantity, int productId)
+        public OrderModel(int shopId, DateTime date, double quantity, int productId)
         {
             ShopId = shopId;
             Date = date;
